Move the db configuration out of `lib`
The configuration is now injected by the `main` "package". One of the consequence is that the model definitions are now wrapped in a function. You can retrieve the models from the API's context.

The next step will be to retrieve the configuration from the environment but it is beyond the scope of this MR.

Signed-off-by: Charles Samborski <demurgos@demurgos.net>